requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless>=4.9.0
numpy>=1.26.0
ultralytics
torch>=2.2.0
torchvision
Pillow==10.1.0
tqdm==4.66.1